requirements.txt
rtl==0.4.3
tweepy==3.9.0
Pillow==7.2.0
